+++ src/test/java/com/gwangya/performance/service/SeatServiceTest.java
@@ -1,8 +1,8 @@
 package com.gwangya.performance.service;
 
-import static com.gwangya.fixture.performance.PerformanceDetailFixture.*;
-import static com.gwangya.fixture.performance.PerformanceFixture.*;
-import static com.gwangya.fixture.performance.SeatFixture.*;
+import static com.gwangya.performance.domain.PerformanceDetailFixture.*;
+import static com.gwangya.performance.domain.PerformanceFixture.*;
+import static com.gwangya.performance.domain.SeatFixture.*;
 import static org.assertj.core.api.Assertions.*;
 
 import java.time.LocalDateTime;
